requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anys==0.3.0
    # via -r requirements.in
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via
    #   cattrs
    #   jsonschema
    #   lsprotocol
    #   referencing
build==1.0.3
    # via pip-tools
cattrs==23.2.3
    # via lsprotocol
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via
    #   pip-tools
    #   pur
coverage[toml]==7.4.0
    # via
    #   coverage
    #   pytest-cov
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
distlib==0.3.8
    # via virtualenv
django==5.0.1
    # via
    #   -r requirements.in
    #   django-autocompletefilter
    #   django-import-export
    #   django-pydantic-field
    #   djangorestframework
    #   drf-orjson-renderer
    #   drf-spectacular
django-autocompletefilter==0.0.8
    # via -r requirements.in
django-import-export==3.3.6
    # via -r requirements.in
django-pydantic-field==0.3.0b2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-orjson-renderer
    #   drf-spectacular
drf-orjson-renderer==1.7.2
    # via -r requirements.in
drf-spectacular==0.27.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==2.0.1
    # via stack-data
filelock==3.13.1
    # via virtualenv
identify==2.5.33
    # via pre-commit
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
ipython==8.20.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jsonschema==4.20.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
lsprotocol==2023.0.0
    # via
    #   pygls
    #   ruff-lsp
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.6
    # via ipython
nodeenv==1.8.0
    # via pre-commit
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
orjson==3.9.10
    # via drf-orjson-renderer
packaging==23.2
    # via
    #   build
    #   pytest
    #   ruff-lsp
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pur==7.3.1
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pydantic==2.5.3
    # via django-pydantic-field
pydantic-core==2.14.6
    # via pydantic
pygls==1.2.1
    # via ruff-lsp
pygments==2.17.2
    # via ipython
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
    #   pytest-unordered
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.7.0
    # via -r requirements.in
pytest-mock==3.12.0
    # via -r requirements.in
pytest-unordered==0.5.2
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   pre-commit
    #   tablib
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
ruff==0.1.13
    # via
    #   -r requirements.in
    #   ruff-lsp
ruff-lsp==0.0.49
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   django-import-export
    #   tablib
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   django-pydantic-field
    #   pydantic
    #   pydantic-core
    #   ruff-lsp
uritemplate==4.1.1
    # via drf-spectacular
virtualenv==20.25.0
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.42.0
    # via pip-tools
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
